feat(activerecord): MySQL/Trilogy adapter connection-adapters layer (#362)
## Summary
- Implements 16 new classes for the MySQL adapter infrastructure in
`connection-adapters/`, matching the Rails MySQL adapter class hierarchy
- Adds `AbstractMysqlAdapter` + `StatementPool`, MySQL-specific
`Column`, `Quoting`, `SchemaCreation`, `SchemaDefinitions`
(TableDefinition, Table, ColumnMethods), `SchemaDumper`,
`SchemaStatements`, `TypeMetadata`, `DatabaseStatements`,
`ExplainPrettyPrinter`
- Adds `Mysql2Adapter` re-export + `Mysql2::DatabaseStatements`,
`TrilogyAdapter` + `Trilogy::DatabaseStatements`
- Moves activerecord API compare from 68.8% to 71.7% (398/555
classes/modules)

## Test plan
- [x] `pnpm run api:compare -- --package activerecord` confirms 16 new
matches (all ✓)
- [x] TypeScript compiles cleanly (`tsc --noEmit` — no new errors)
- [x] ESLint + Prettier pass via pre-commit hooks

+++ packages/activerecord/src/base.ts
@@ -356,7 +356,7 @@ export class Base extends Model {
       const pkDef = isPg
         ? `${quoteIdentifier(pk, adapterName)} SERIAL PRIMARY KEY`
         : isMysql
-          ? `${quoteIdentifier(pk, adapterName)} INT AUTO_INCREMENT PRIMARY KEY`
+          ? `${quoteIdentifier(pk, adapterName)} BIGINT AUTO_INCREMENT PRIMARY KEY`
           : `${quoteIdentifier(pk, adapterName)} INTEGER PRIMARY KEY AUTOINCREMENT`;
       colDefs.push(pkDef);
     } else {
